requirements.txt
phidata
groq
yfinance
openai
packaging
duckduckgo-search
phidata
sqlalchemy
fastapi[standard]
streamlit
gradio